backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-jose[cryptography]
passlib
python-multipart